requirements.txt
contourpy==1.1.1
cycler==0.12.1
fonttools==4.50.0
importlib-resources==6.3.1
kiwisolver==1.4.5
matplotlib==3.7.5
numpy==1.24.4
packaging==24.0
pillow==10.2.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
six==1.16.0
zipp==3.18.1
